+++ spec/features/universe_factions_spec.rb
@@ -16,7 +16,7 @@ describe "Get factions" do
   specify do
     expect(subject.factions.first.as_json).to eq(corporation_id: 1_000_084,
       description: "The largest of the five main empires, the Amarr Empire is a sprawling patch-work of feudal-like provinces held together by the might of the emperor. Religion has always played a big part in Amarrian politics and the Amarrians believe they are the rightful masters of the world, souring their relations with their neighbours. Another source of ill-feelings on part of the other empires is the fact that the Amarrians embrace slavery.",
-      faction_id: 500_003,
+      id: 500_003,
       is_unique: true,
       militia_corporation_id: 1_000_179,
       name: "Amarr Empire",