+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 * Drop Rails 7.1
 * Fix POST requests
 * Upgrade "/v2/characters/{character_id}/blueprints/" to "/v3/characters/{character_id}/blueprints/"
+* Rename `EveOnline::ESI::Models::Ancestry#ancestry_id` to `#id`
+* Rename `EveOnline::ESI::Models::Bloodline#bloodline_id` to `#id`
+* Rename `EveOnline::ESI::Models::Race#race_id` to `#id`
+* Upgrade "/v5/characters/%<character_id>s/notifications/" to "/v6/characters/%<character_id>s/notifications/"
 
 ## v0.46.0