+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Rename `EveOnline::ESI::Models::Ancestry#ancestry_id` to `#id`
 * Rename `EveOnline::ESI::Models::Bloodline#bloodline_id` to `#id`
 * Rename `EveOnline::ESI::Models::Race#race_id` to `#id`
+* Rename `EveOnline::ESI::Models::Category#category_id` to `#id`
 * Upgrade `/v5/characters/%<character_id>s/notifications/` to `/v6/characters/%<character_id>s/notifications/`
 * Upgrade `/v1/status/` to `/v2/status/`