+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Ruby 3.4
 * Drop Rails 7.1
 * Fix POST requests
+* Add `faraday-http-cache` gem as dependency
 * Upgrade `/v2/characters/{character_id}/blueprints/` to `/v3/characters/{character_id}/blueprints/`
 * Rename `EveOnline::ESI::Models::Ancestry#ancestry_id` to `#id`
 * Rename `EveOnline::ESI::Models::Bloodline#bloodline_id` to `#id`